Move `Timestamps` enum to the timestamps mod
