+++ src/migration/timestamp/mod.rs
@@ -5,12 +5,36 @@ use sea_orm_migration::prelude::*;
 
 pub mod m20240723_201404_add_update_timestamp_function;
 
+/// Timestamp related fields.
+#[derive(DeriveIden)]
+#[non_exhaustive]
+pub enum Timestamps {
+    /// When the row was created. When used with the [crate::migration::schema::timestamps] method, will default to
+    /// the current timestamp (with timezone).
+    CreatedAt,
+    /// When the row was updated. When used with the [crate::migration::schema::timestamps] method, will be initially set to
+    /// the current timestamp (with timezone).
+    ///
+    /// To automatically update the value for a row whenever the row is updated, include the
+    /// [m20240723_201404_add_update_timestamp_function::Migration]
+    /// in your [MigratorTrait] implementation, along with a [MigrationTrait] for your table
+    /// that add a trigger to update the column. Helper methods are provided for this in
+    /// the [crate::migration::timestamp] module. Specifically, see:
+    /// - [exec_create_update_timestamp_trigger]
+    /// - [exec_drop_update_timestamp_trigger]
+    ///
+    /// Note that the auto-updates mentioned above are currently only supported on Postgres. If
+    /// an app is using a different DB, it will need to manually update the timestamp when updating
+    /// a row.
+    UpdatedAt,
+}
+
 /// Wrapper around [create_update_timestamp_function] to execute the returned [Statement], if
 /// present.
 ///
 /// # Examples
 /// ```rust
-/// use roadster::migration::schema::Timestamps;
+/// use roadster::migration::timestamp::Timestamps;
 /// use roadster::migration::timestamp::exec_create_update_timestamp_function;
 /// use sea_orm_migration::prelude::*;
 ///
@@ -47,7 +71,7 @@ pub async fn exec_create_update_timestamp_function<C: IntoIden>(
 ///
 /// # Examples
 /// ```rust
-/// use roadster::migration::schema::Timestamps;
+/// use roadster::migration::timestamp::Timestamps;
 /// use roadster::migration::timestamp::{exec_create_update_timestamp_function_dep_column};
 /// use sea_orm_migration::prelude::*;
 ///
@@ -204,7 +228,7 @@ $$ language 'plpgsql';
 ///
 /// # Examples
 /// ```rust
-/// use roadster::migration::schema::Timestamps;
+/// use roadster::migration::timestamp::Timestamps;
 /// use roadster::migration::timestamp::exec_drop_update_timestamp_function;
 /// use sea_orm_migration::prelude::*;
 ///
@@ -269,7 +293,7 @@ fn drop_update_timestamp_function_for_db_backend<C: IntoIden>(
 ///
 /// # Examples
 /// ```rust
-/// use roadster::migration::schema::Timestamps;
+/// use roadster::migration::timestamp::Timestamps;
 /// use roadster::migration::timestamp::exec_create_update_timestamp_trigger;
 /// use sea_orm_migration::prelude::*;
 ///
@@ -359,7 +383,7 @@ EXECUTE PROCEDURE {fn_call};
 ///
 /// # Examples
 /// ```rust
-/// use roadster::migration::schema::Timestamps;
+/// use roadster::migration::timestamp::Timestamps;
 /// use roadster::migration::timestamp::exec_drop_update_timestamp_trigger;
 /// use sea_orm_migration::prelude::*;
 ///
