+++ src/migration/user/m20240724_005115_user_update_timestamp.rs
@@ -6,7 +6,7 @@
 //!
 //! Note: Currently only supports Postgres. If another DB is used, will do nothing.
 
-use crate::migration::schema::Timestamps;
+use crate::migration::timestamp::Timestamps;
 use crate::migration::timestamp::{
     exec_create_update_timestamp_trigger, exec_drop_update_timestamp_trigger,
 };
